Move events to logs API
